# lib/api-zod/src/generated/types/roleSuggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoleSuggestion {
  suggestion: string;
  category: string;
  emoji: string;
}